Fix SecondVaccine issue
In the data extract, the number of NULLs for Vaccine1Date and Vaccine2Date were exactly the same. I have fixed the code to allow nulls in SecondVaccineDate and this  has solved the issue. There are now some patients with a first vaccine but not a second which makes sense.